refactor: retrieval과 reranking 로직을 application layer로 승격
