+++ src/main/java/com/techfork/domain/recommendation/listener/PersonalizedProfileGeneratedEventListener.java
@@ -1,6 +1,6 @@
 package com.techfork.domain.recommendation.listener;
 
-import com.techfork.domain.recommendation.service.RecommendationService;
+import com.techfork.recommendation.application.generation.RecommendationService;
 import com.techfork.personalization.application.event.PersonalizedProfileGeneratedEvent;
 import com.techfork.useraccount.application.query.lookup.UserLookupService;
 import com.techfork.useraccount.domain.User;
