+++ src/test/java/com/techfork/domain/recommendation/service/RecommendationCommandServiceTest.java
@@ -1,5 +1,6 @@
 package com.techfork.domain.recommendation.service;
 
+import com.techfork.recommendation.application.generation.RecommendationService;
 import com.techfork.useraccount.application.query.lookup.UserLookupService;
 import com.techfork.useraccount.domain.User;
 import org.junit.jupiter.api.DisplayName;
