+++ src/test/java/com/techfork/evaluation/recommendation/RecommendationEvaluationService.java
@@ -11,10 +11,10 @@ import com.techfork.recommendation.config.RecommendationProperties;
 import com.techfork.recommendation.domain.RecommendedPost;
 import com.techfork.recommendation.infrastructure.RecommendationHistoryRepository;
 import com.techfork.recommendation.infrastructure.RecommendedPostRepository;
-import com.techfork.domain.recommendation.service.LlmRecommendationService;
-import com.techfork.domain.recommendation.service.MmrService;
-import com.techfork.domain.recommendation.service.MmrService.MmrCandidate;
-import com.techfork.domain.recommendation.service.MmrService.MmrResult;
+import com.techfork.recommendation.application.generation.LlmRecommendationService;
+import com.techfork.recommendation.application.ranking.MmrService;
+import com.techfork.recommendation.application.ranking.MmrService.MmrCandidate;
+import com.techfork.recommendation.application.ranking.MmrService.MmrResult;
 import com.techfork.global.util.RrfScorer;
 import com.techfork.personalization.infrastructure.PersonalizationProfileDocument;
 import com.techfork.personalization.application.query.lookup.PersonalizationProfileLookupService;
