+++ src/test/java/com/techfork/personalization/integration/PersonalizedProfileGeneratedAfterCommitIntegrationTest.java
@@ -1,7 +1,7 @@
 package com.techfork.personalization.integration;
 
 import com.techfork.domain.recommendation.listener.PersonalizedProfileGeneratedEventListener;
-import com.techfork.domain.recommendation.service.RecommendationService;
+import com.techfork.recommendation.application.generation.RecommendationService;
 import com.techfork.personalization.application.event.PersonalizedProfileGeneratedEvent;
 import com.techfork.useraccount.application.query.lookup.UserLookupService;
 import com.techfork.useraccount.domain.User;
